出品管理機能（単体テストコード～jsを使用した計算結果成功）
