+++ app/models/item.rb
@@ -11,11 +11,10 @@ class Item < ApplicationRecord
   belongs_to_active_hash :shipping_day
   
   #空の投稿を保存できないようにする
-  validates :image, :item_name, :item_info, :price, :category, :prefecture_id,  presence: true
+  validates :image, :item_name, :item_info, :price, :category_id, :prefecture_id,  presence: true
   #ジャンルの選択が「---」の時は保存できないようにする
   validates :category_id, :condition_id, :shipping_cost_id, :prefecture_id, :shipping_day_id, presence: true
-
   
-  validates :price, numericality: { only_integer: true }
+  validates :price, numericality: { only_integer: true , greater_than_or_equal_to: 300, less_than_or_equal_to: 9999999 }
 
 end
