+++ config/routes.rb
@@ -3,8 +3,8 @@ Rails.application.routes.draw do
   devise_for :users
   
   root to: 'items#index'
+  
   resources :items, only: [:index, :new, :create]
-
 
   resources :articles, only: [:new]
 end
