+++ spec/fixtures/test_image.jpgZone.Identifier
@@ -1,0 +1,4 @@
+[ZoneTransfer]
+ZoneId=3
+ReferrerUrl=https://www.gettyimages.co.jp/
+HostUrl=https://media.gettyimages.com/id/977798706/ja/%E3%82%B9%E3%83%88%E3%83%83%E3%82%AF%E3%83%95%E3%82%A9%E3%83%88/red-fox-on-meadow-in-amsterdamse-waterleidingduinen-nature-reserve-netherlands.jpg?s=1024x1024&w=gi&k=20&c=NegiQ-rF_KzX-x_y6rNU0iY6SbgvFYiFEKTmdYy22zA=
