+++ spec/models/item_spec.rb
@@ -1,5 +1,86 @@
 require 'rails_helper'
 
 RSpec.describe Item, type: :model do
-  pending "add some examples to (or delete) #{__FILE__}"
+  before do
+    @item = FactoryBot.build(:item)
+  end
+
+describe '商品出品' do
+    context '商品出品がうまくいくとき' do
+      it '全ての項目が存在すれば出品できる' do
+        expect(@item).to be_valid
+      end
+    end
+
+    context '商品出品がうまくいかないとき' do
+      it '画像が空では登録できない' do
+        @item.image = nil
+        @item.valid?
+        expect(@item.errors.full_messages).to include("Image can't be blank")
+      end
+
+      it '商品名が空では登録できない' do
+        @item.item_name = nil
+        @item.valid?
+        expect(@item.errors.full_messages).to include("Item name can't be blank")
+      end
+
+      it '商品説明が空では登録できない' do
+        @item.item_info = nil
+        @item.valid?
+        expect(@item.errors.full_messages).to include("Item info can't be blank")
+      end
+
+      it 'カテゴリ情報がないと登録できない' do
+        @item.category_id = nil
+        @item.valid?
+        expect(@item.errors.full_messages).to include("Category can't be blank")
+      end
+
+      it '商品の状態情報が空では登録できない' do
+        @item.condition_id = nil
+        @item.valid?
+        expect(@item.errors.full_messages).to include("Condition can't be blank")
+      end
+
+      it '配送料負担情報が空では登録できない' do
+        @item.shipping_cost_id = nil
+        @item.valid?
+        expect(@item.errors.full_messages).to include("Shipping cost can't be blank")
+      end
+
+      it '発送元地域が空では登録できない' do
+        @item.prefecture_id = nil
+        @item.valid?
+        expect(@item.errors.full_messages).to include("Prefecture can't be blank")
+      end
+    
+      it '発送までの日数情報が空では登録できない' do
+        @item.shipping_day_id = nil
+        @item.valid?
+        expect(@item.errors.full_messages).to include("Shipping day can't be blank")
+      end
+
+      it '価格が空では登録できない' do
+        @item.price = nil
+        @item.valid?
+        expect(@item.errors.full_messages).to include("Price can't be blank")
+      end
+
+      it '価格が¥300~¥9,999,999の範囲外だと登録できない' do
+        @item.price = 299
+        @item.valid?
+        expect(@item.errors.full_messages).to include("Price must be greater than or equal to 300")
+        @item.price = 10_000_000
+        @item.valid?
+        expect(@item.errors.full_messages).to include("Price must be less than or equal to 9999999")
+      end
+
+      it '価格が半角数字のみで保存可能であることを確認する' do
+        @item.price = '３００'
+        @item.valid?
+        expect(@item.errors.full_messages).to include('Price is not a number')
+      end
+    end
+  end
 end
